Updated program verbage to better relay that all calculations are estimations.
Streamlined calculations resulting in better accuray for calculations by basing most calculations on the 52 weeks in the year. This results in a significant accuracy boost for yearly and monthly considerations in calculations. No real changes for weekly/biweekly calculations.

Added confirmation loops for user input to cut down on entry errors resulting in incorrect calculations.

Cleaned up code to improve the workflow of the program and to improve interface flow aesthetics for the user.

Broke the calculator for annual wages in the process.